requirements.txt
streamlit
rembg
pillow
opencv-python-headless
numpy
onnxruntime